Split off common 2D pspec code into a base class
